Get element names from external service
Summary:
This implements retrieval of element names from an external service (https://data.met.no/elements) when running in prod mode.

This has to be reviewed in conjunction with D788.

Test Plan:
- `activator clean test`
- `activator run` and test the Mock framework.

The production framework cannot be tested without a working /elements service, which requires a metapi instance with /elements (which is provided by D788).

Reviewers: joa, #met_api, borgem

Reviewed By: #met_api, borgem

Subscribers: borgem

Maniphest Tasks: T853

Differential Revision: https://phab.met.no/D778