Fixed elements authorization
Summary:
The shipped version of metapi-observations did not include proper
authentication in the data retrieval from /elements. This was
unfortunately all too easy to miss in Dev mode, and didn't get
picked up in CD since there are no acceptance tests.

This patch picks up the Authorization header from the observations
request and passes it on to the /elements request.

Test Plan:
As always, check coverage: `activator clean coverage test`

One way to check this works properly:
- `activator clean stage dist`
- Make sure your metapi is up to date, and set up correctly with
`auth.active=true` and the kdvh and elements databases configured.
- Copy target/scala-2.11/observations_2.11-0.2-SNAPSHOT.jar to metapi/lib
- In metapi's build.sbt, comment out the dependency on observations

As an alternative to the above, do a `activator clean compile publishLocal`;
but be careful that you have cleaned out other libraries before this.

- activator 'testProd -Dplay.evolutions.db.authorization.autoApply=true'

This should spin up a test production instance. Run a query such as:
http://localhost:9000/observations/v0.jsonld?sources=KN18700&reftime=2015-05-01T12%3A00%3A00&elements=air_temperature

And verify that you are actually receiving data.

Reviewers: #met_api, geira

Reviewed By: #met_api, geira

Subscribers: geira

Maniphest Tasks: T1598

Differential Revision: https://phab.met.no/D805