Avoided 414 URI Too Long by issuing multiple GET requests
When internally querying the elements/ endpoint for a large number of elements (typically the case when
leaving the 'elements' query parameter empty in the original request), the GET request may fail with
414 URI Too Long. This fix works around the problem by breaking the GET request into multiple smaller
ones.

The ideal solution would be to issue a single POST request rather than multiple GET requests, since we
then wouldn't have to deal with the problem of finding an optimal subrequest size, and would probably also
achieve a faster overall response. But this would require support for POST requests in the elements/ endpoint.