Don't fail with 404 on unavailable src/elem combos
A request for multiple combinations of source and element,
like '...sources=src1,src2...&elements=elem1,elem2...' would result
in overall 404 Not found failure if data was unavailable for
at least one of the source/element combinations.

For such cases, this fix ensures the expected behavior of returning
status=200 and silently leaving out the missing data in the output.