Supported value tables with tdato column
Certain elements are associated with a value table (t_elem_map_timeseries.table_name)
that contains a tdato column (in addition to dato). This change ensures that the
query for such elements are filtered correctly based on the period encoded in the
element name. For example, the element best_estimate_mean(air_temperature P1M) will
only match rows in the value table where the difference tdato-dato is around 30 days.